+++ test/fixtures/begin.rb
@@ -5,3 +5,7 @@ end
 begin
   expression
 end
+%
+case value
+in ^(expression)
+end