+++ test/fixtures/var_field.rb
@@ -8,3 +8,13 @@ $foo = bar
 foo = bar
 %
 @foo = bar
+%
+foo in bar
+%
+foo in ^bar
+%
+foo in ^@bar
+%
+foo in ^@@bar
+%
+foo in ^$gvar